pyproject.toml
[project]
name = "llm-serving"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hf-xet>=1.2.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.1",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.6",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.9",
    "protobuf>=6.33.2",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "pyngrok>=7.5.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.0",
    "sentencepiece>=0.2.1",
    "streamlit>=1.52.2",
    "transformers>=4.57.3",
    "unstructured>=0.18.26",
]
